+++ src/app/docs/layout/demo-drag-n-drop/demo-drag-n-drop.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
     selector: 'app-demo-drag-n-drop',
     templateUrl: './demo-drag-n-drop.component.html',
-    styleUrls: ['./demo-drag-n-drop.component.scss']
+    styleUrls: ['./demo-drag-n-drop.component.scss', '../../demo.module.scss']
 })
 export class DemoDragNDropComponent implements OnInit {