+++ src/app/docs/modals/demo-popup/demo-popup.component.ts
@@ -1,27 +1,90 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { AlertService, PopupService } from 'ngx-slice-kit';
+import { DemoPageModel } from '../../../shared/model';
 
 @Component({
     selector: 'app-demo-popup',
     templateUrl: './demo-popup.component.html',
-    styleUrls: ['./demo-popup.component.scss']
+    styleUrls: ['../../demo.module.scss']
 })
 export class DemoPopupComponent implements OnInit {
 
+    @ViewChild('defaultRef', {static: true}) defaultRef: any;
+
+    page: DemoPageModel;
+
     constructor(
         private popup: PopupService,
-        private alert: AlertService
+        private alert: AlertService,
     ) {
     }
 
-    ngOnInit(): void {
-    }
-
-    openPopup(): void {
+    showPopup(): void {
         this.popup.showPopup({
             message: `Can contain\nmultiple strings\nseparated by \\n`
         }).subscribe(ok => {
             ok ? this.alert.success({message: 'Success'}) : this.alert.error({message: 'Unsuccess'});
         });
     }
+
+    ngOnInit(): void {
+        this.page = {
+            title: 'Popup component example',
+            subtitle: '',
+            demos: [
+                {
+                    title: 'Default usage',
+                    description: '',
+                    templateRef: this.defaultRef,
+                    values: {
+                        html: `<button (click)="showPopup()" sdk-raised-button>Open example popup</button>`,
+                        module: `import { PopupModule } from 'ngx-slice-kit';
+
+@NgModule({
+    imports: [
+        CommonModule,
+        FormsModule,
+        // add PopupModule to app imports
+        PopupModule,
+    ],
+})
+export class DemoPopupModule {
+}`,
+                        component: `import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-demo-popup',
+    templateUrl: './demo-popup.component.html',
+    styleUrls: ['./demo-popup.component.scss']
+})
+export class DemoPopupComponent {
+
+    constructor(
+        private popup: PopupService,
+        private alert: AlertService,
+    ) {
+    }
+
+    showPopup(): void {
+        this.popup.showPopup({
+            message: \`Can contain\\nmultiple strings\\nseparated by \\\\n\`
+        }).subscribe(ok => {
+            ok ? this.alert.success({message: 'Success'}) : this.alert.error({message: 'Unsuccess'});
+        });
+    }
+
+}`,
+                    },
+                }
+            ],
+            apis: [
+                {
+                    label: '',
+                    type: '',
+                    description: '',
+                }
+            ],
+        };
+    }
+
 }