+++ src/app/docs/modals/modals.module.ts
@@ -13,6 +13,7 @@ import { DemoDialogPreview2Component } from './demo-dialog/demo-dialog-preview2/
 import { DemoPopoverComponent } from './demo-popover/demo-popover.component';
 import { DemoPopupComponent } from './demo-popup/demo-popup.component';
 import { DemoTooltipComponent } from './demo-tooltip/demo-tooltip.component';
+import { SharedModule } from "../../shared/shared.module";
 
 const entryComponents = [];
 
@@ -33,6 +34,7 @@ const moduleDeclarations = [
         CommonModule,
         DemoModalsRoutingModule,
         SliceKitModule,
+        SharedModule,
     ],
     declarations: moduleDeclarations,
     exports: moduleDeclarations,